+++ src/main/java/core/di/factory/bean/BeanInvokeType.java
@@ -1,4 +1,4 @@
-package core.di.factory;
+package core.di.factory.bean;
 
 /**
  * Created By kjs4395 on 7/20/20
