+++ src/main/java/core/di/factory/bean/BeanMaker.java
@@ -1,0 +1,13 @@
+package core.di.factory.bean;
+
+import core.di.factory.BeanFactory;
+
+/**
+ * Created By kjs4395 on 7/20/20
+ */
+public interface BeanMaker {
+     boolean isSupport(BeanInfo beanInfo);
+
+     <T> T makeBean(BeanInfo beanInfo, BeanFactory beanFactory);
+
+}
