+++ src/test/java/core/di/factory/BeanFactoryTest.java
@@ -4,7 +4,8 @@ import com.google.common.collect.Sets;
 import core.annotation.Repository;
 import core.annotation.Service;
 import core.annotation.web.Controller;
-import core.di.ConfigurationScanner;
+import core.di.factory.bean.BeanInfo;
+import core.di.factory.bean.BeanInvokeType;
 import core.di.factory.example.MyQnaService;
 import core.di.factory.example.QnaController;
 import org.junit.jupiter.api.BeforeEach;
@@ -19,17 +20,15 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class BeanFactoryTest {
     private Reflections reflections;
-    private BeanFactory2 beanFactory;
+    private BeanFactory beanFactory;
 
     @BeforeEach
     @SuppressWarnings("unchecked")
     public void setup() {
-        ConfigurationScanner configurationScanner = new ConfigurationScanner();
-        configurationScanner.initialize();
 
         reflections = new Reflections("core.di.factory.example");
         Set<Class<?>> preInstanticateClazz = getTypesAnnotatedWith(Controller.class, Service.class, Repository.class);
-        beanFactory = new BeanFactory2();
+        beanFactory = new BeanFactory();
         beanFactory.register(getBeanInfos(preInstanticateClazz));
         beanFactory.initialize();
     }
