+++ src/test/java/core/di/factory/ConfigurationScannerTest.java
@@ -5,6 +5,8 @@ import core.di.factory.example.ExampleConfig;
 import core.di.factory.example.IntegrationConfig;
 import core.di.factory.example.JdbcUserRepository;
 import core.di.factory.example.MyJdbcTemplate;
+import core.di.factory.scanner.ClasspathBeanScanner;
+import core.di.factory.scanner.ConfigurationBeanScanner;
 import org.junit.jupiter.api.Test;
 
 import javax.sql.DataSource;
@@ -19,8 +21,8 @@ public class ConfigurationScannerTest {
     @Test
     public void configurationScannerTest() {
 
-        BeanFactory2 beanFactory = new BeanFactory2();
-        ConfigurationBeanScanner2 cbs = new ConfigurationBeanScanner2(beanFactory);
+        BeanFactory beanFactory = new BeanFactory();
+        ConfigurationBeanScanner cbs = new ConfigurationBeanScanner(beanFactory);
         cbs.register(ExampleConfig.class);
         beanFactory.initialize();
 
@@ -29,8 +31,8 @@ public class ConfigurationScannerTest {
 
     @Test
     public void register_classpathBeanScanner_통합() {
-        BeanFactory2 beanFactory = new BeanFactory2();
-        ConfigurationBeanScanner2 cbs = new ConfigurationBeanScanner2(beanFactory);
+        BeanFactory beanFactory = new BeanFactory();
+        ConfigurationBeanScanner cbs = new ConfigurationBeanScanner(beanFactory);
         cbs.register(IntegrationConfig.class);
         beanFactory.initialize();
